Rename package to stockpaydividend
