+++ stockpaydividend-TokenSDK/TRADEMARK
