+++ stockpaydividend-TokenSDK/contracts/build.gradle
