+++ stockpaydividend-TokenSDK/gradle.properties
