+++ stockpaydividend-TokenSDK/settings.gradle
